move common functions to fin.hpp
